demos/web3_py_simple_storage/requirements.txt
py-solc-x
web3
black
eth-brownie